+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Fixes
 
+- Improved error messages on Android for failed `confirmPayment` and `confirmSetupIntent` calls, and any Google Pay related methods. [#957](https://github.com/stripe/stripe-react-native/pull/957)
+
 ## 0.10.0
 
 ### Breaking changes